+++ elenasrc3/engine/windows/winroutines.cpp
@@ -62,7 +62,7 @@ uintptr_t SystemRoutineProvider :: ExpandPerm(void* allocPtr, size_t newSize)
    return !r ? 0 : (uintptr_t)allocPtr;
 }
 
-void* SystemRoutineProvider :: CreateThread(int tt_index, int flags, void* threadProc)
+void* SystemRoutineProvider :: CreateThread(size_t tt_index, int flags, void* threadProc)
 {
    return ::CreateThread(
       nullptr,                            // default security attributes
